fix search terms query
